Always perform multi key lookups atomically
When we perform a multi key lookup, for example for
a `WHERE IN (?, ?, ?)` query, we may miss on some keys
and end up with results that represent the map at
different time points. This can lead to pretty inconsistent
results, and even if no misses happen lookups happen
non atomically, and a publish can happen in between lookups.

This change modifies all reads to perform atomic lookups,
under the same read guard into the map.

It also slightly changes the behavior of `remove_range` and
`empty_range`, to accept owned BoundPair, as opposed to a
RangeBounds that are always cloned.

Fixes: ENG-1449
Change-Id: I9b816398ff800d7d68e64f145b7012decb97a7b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2683
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>