Share server options between server and adapter
Now that we're going forward with more-permanently being able to embed a
server worker inside an adapter process, we'd like to be able to
consistently set *all* server worker flags when running the adapter, and
not have to manually expose them like we did for the experimental query
flags and replication SSL verification disabling. To that end, this
commit breaks out a new `noria_server::WorkerOptions` clap options
struct from the Options in noria_server's main.rs, adds a method to
create a noria_server::Builder from that, and uses it via
`#[clap(flatten)]` in both noria_server's and noria_client_adapter's
top-level Options struct.

Fixes: ENG-1439
Change-Id: I275d56515c844db70278f23239cd67fade33ebba
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2657
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>