readyset-cloud: Initialize cloud monorepo with TS.
Using pnpm alone and no other monorepo tooling as of yet because of
various limitations of those different managers.

The current data plane is just static structs to show how it will work
and the next phase will be about replacing this with something that uses
JSON files nearby with validation.

The CDK usage in the control plane is also extremely minimal consisting
of only setting up providers. These constructs will be replaced by the
currently in progress work.

This also includes documentation of the general structure that is being
planned for this system.

This has also not yet been integrated into CI and will be coming soon.

Change-Id: I93d7be56a5e6712856dccfc813e4b8a99da04b7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2667
Tested-by: Buildkite CI
Reviewed-by: Marcus Volhontseff <marcus.v@readyset.io>