mir: Fix off-by-one error in columns impl for JoinAggregates
The previous implementation of column deduplication for JoinAggregates
wasn't just ugly, it also had an off-by-one error where if the last
column in the list was removed we wouldn't step backwards, causing us to
panic with an index error on the next iteration of the loop. This
replaces the whole thing with a fold which is a lot cleaner and not
prone to these kinds of issues.

Refs: ENG-1427
Change-Id: Iad3363fdc2727fc782d0e198ad3dead8363fa50b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2669
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>