query_graph: Don't add aggregates from ORDER BY to expr columns
OutputColumn::Expr in particular is supposed to be used for expressions
which are expected to be projected by a Project node, but Project
nodes (obviously) don't have the ability to evaluate aggregates. We're
already adding the aggregate to the `aggregates` of the query graph,
which already suffices to ensure the aggregate is evaluated for us, and
then the pull_columns pass will pull it through if necessary (though it
might not end up being necessary).

Fixes: ENG-1427
Change-Id: I29b024fcaee397834a179357e643e6c24960b50e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2670
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>