Implement k-way merge via tournament trees
This change adds a new crate to the monorepo, called
tournament-kway, this crate implements a StreamingIterator
that can merge other, sorted, StreamingIterators, such that
the result is also sorted. The merging is performed lazily
so we can efficienly take top n elements from a list of
sorted arrays. This crate is needed for a future functionality
where we will be mergin the result sets of different keys.

Change-Id: Id7b69187ac892b4bc304b47d1d613f98414cc54f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2671
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI