query_graph: Dedupe aggregates in queries
As an optimization, avoid creating duplicates of the same aggregate node
in a query, either if the aggregate is explicitly projected twice in the
query source (including with different alias) or if the aggregate is
both projected in the column list and eg referenced in the ORDER BY
clause (the case that caused me to make this change). This is
accomplished by storing aggregates in QueryGraph as a
`HashMap<FunctionExpr, SqlIdentifier>` instead of just a vec of tuples,
and in the case that an aggregate is referenced a second time with a
different alias the `OutputColumn::Data` we make just references the
alias of the first aggregate.

Since QueryGraph needs to be able to semi-consistently implement Hash,
and does so by ordering the various lists it stores internally, this
also required deriving PartialOrd for a significant portion of the
nom-sql AST to allow us to order those FunctionExpr terms.

Change-Id: I468f530bf6688eab440d938db9e77128a2e6be6b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2673
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>