Ignore sharded tests, unshard non-sharded tests
It seems that ignoring the `infers_type_for_grouped` test
only reduced the occurnace of hanging CI, not entirely solved
it. I digged a little more, and running all server tests in
a loop does still hang occasionally on various tests.

I found out that what is common for all of those tests is
that they are stuck on a "lookup", and that they are sharded.

Although this CL does not solve the problem whatsoever, it
does ignore all the shareded tests for now, while turning
all tests that are not explicitly sharded, to unsharded.

This also unignores `infers_type_for_grouped` so that it
is easier to see if the problem is indeed solved by this
change.

Refs: ENG-1438
Change-Id: Ia53a2e5a6ee3753020474ba3ce13692aa28e2356
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2674
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI