noria-client: Fix leaving proxy mode on transaction end
Because of the order of the match arms here, we'd end up just proxying
the COMMIT/ROLLBACK upstream unconditionally, without inspecting it to
determine that it meant we should leave ProxyMode::InTransaction. This
adds a special-case match arm for SqlQuery::Commit before we check
proxy_state.should_proxy - this is a little messy, but feels like the
cleanest way to accomplish things with the way stuff is currently
structured.

Also, adds an integration test to check that committing a transaction
allows us to start caching queries again.

Fixes: ENG-1448
Change-Id: If89ee0e116812b5ec62e06848ee1547b3dadc5d2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2678
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>