nom-sql: Parse query with OFFSET but without LIMIT
PostgreSQL (but not MySQL) allows SELECT statements to specify an OFFSET
clause without a LIMIT clause, and Prisma appears to do this for all
queries (even those without pagination, where it just specifies OFFSET
0). Even if we don't support non-zero offset here we should be able to
parse this and our AST should be able to handle it, if only to support 0
offset.

Since making both Offset and Limit optional makes the LimitClause struct
kind of pointless, this also removes that and just makes `limit` and
`offset` fields of both SelectStatement and CompoundSelectStatement.

Besides all of that, this doesn't actually add any new query *support* -
everywhere we encounter an offset without a limit we just return an
error, at least for now.

Fixes: ENG-1453
Change-Id: I43835a656de4d4e69a3d15fece5bd1c923073109
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2679
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>