readyset-cloud: Initial data plane
This uses an in-repository data plane. In addition, there are schemas
and validation as part of the process of processing the files. If the
tests pass, then the files have all been parsed and validated which
should prevent against various issues.

These files are not being put in a database for now as this is slightly
easier to evolve the schema and does not require additional locking
around who is currently editing things like a database would require.

Change-Id: I1fa241b18a08e6fa878c51fd297b940333708f38
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2680
Tested-by: Buildkite CI
Reviewed-by: Marcus Volhontseff <marcus.v@readyset.io>
Reviewed-by: Ian Meyer <ian@readyset.io>