noria-client: Support bare OFFSET ? if ? = 0
We can't support arbitrary parametrized OFFSET without a LIMIT in the
general case, but the specific case of `OFFSET 0` is a no-op, and we'd
like to be able to detect that case since Prisma appears to send it for
all of their queries. This adds a pass to `noria_client/rewrite` that
strips off all "bare" (without a corresponding LIMIT) parametrized
offsets, then when building keys checks that the value supplied to that
parameter is 0, returning an unsupported error if it isn't and just
removing the parameter entirely if it is.

Fixes: ENG-1454
Change-Id: Ifae29ad95794128f13b7019d3046d930f5e95b74
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2684
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>