Add support for ARRAY literals
Add a new `Literal::Array` literal variant to `nom_sql::Literal`, and
add support for parsing it and converting it to a DataType.

To keep with the general "the AST implements syntax, not semantics"
theme, all shape checks etc are deferred to the actual DataType
construction, rather than happening at parse time.

Fixes: ENG-1412
Change-Id: I0150ac6cbd110cd9106e1bf44badeaa3e030ebbd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2686
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>