replicators: Support table filtering in MySQL replication
This change introduces the option to only replicate specific
tables from MySQL, by adding the -replication-tables command
line option.

This change also introduces namespace support in replication,
the replicator can now replicate across multiple MySQL databases
on the same server.

By default all tables in the connected database are replicated,
but if the -replication-tables option is specified then only
the named tables will be replicated.

The argument expects a list of comma separated tables in MySQL
format, which can be either the name of the table in the
connected database, a fully qualified name for a table in
another database, or the * wildcard to replicated all tables
in a given database.

For example: -replication-tables tbl1,db1.tbl1,\`.\`.\`,\`,db2.*

Note however that this change does not yet have a way to let
noria know what namespace a table belongs to, that work is
tracked in ENG-1447.

Fixes: ENG-1422
Change-Id: I3f9aacee4973289a992cc169276163e772c01f3c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2688
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>