replicators: Support table filtering in PostgreSQL replication
Similar to MySQL, this change will filter out table and ddl
actions for unwanted tables.

PostgreSQL supports replication filtering natively via explicit
naming in the replication publication slot. However if explicit
names are used, then new tables are not replicated, and that is
something that we *do* want to enable (otherwise why replicate
CREATE TABLE?).

In addition it also changes the visibility of some methods
and structs that were not supposed to be pub to pub(crate).

Fixes: ENG-1423
Change-Id: Ie8a0bffcf32192b3842e8af59fa3a1ccaa7343fb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2692
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>