Separate Autocommit as discrete state
This CL separates autocommit out into it's own state, and handles them
separately from other unsupported sets. This allows us to turn proxying
on and off accordingly. For instance, if we receive `SET autocommit=0`
then we will proxy all queries until we receive a `SET autocommit=1`,
after which we will return to our normal workflow.

This CL also ensures that we don't overwrite the ProxyAlways or Never
states. In the case that we receive a SET that is unsupported, which is
not autocommit related, that should cause a ProxyAlways state to
persist, regardless of whether autocommit is turned back on or not.

Change-Id: I49c1c9912ac2d94413ecc9d1a5cfc5e755db9d0c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2695
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>