docs: Update documentation for GraphViz key count features
Now that we have key counts in GraphViz, the docs need to be updated to
explain what the new numbers in the graph represent. I've also updated
the example graph to match the new output format. The graph structure is
not exactly the same as the current graph, because newer versions of
ReadySet appear to create some extra project_reorder nodes that didn't
used to be introduced, but it's very close to the same other than that.

Change-Id: Ic4f58b0c4b7f3a9ed25d451b8fb6c6709d973d3b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2697
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI