Don't Panic when Replicating Custom Types
Attempting to replicate tables with custom types would cause us to panic
because of a call to Type::from_oid(oid).unwrap(). Type::from_oid()
returns None if called with an oid that does not correspond to a
postgres type.

The error type of a few functions is changed from tokio_postgres::Error
to ReadySetError so that we don't have to create a new enum error type.

Change-Id: I14e5625438834d147f2425da24d3c99cc0ce91ca
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2698
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Tested-by: Buildkite CI