Allow ViewDestroyed queries to retry via Noria
Prior to this commit, if a query received a ViewDestroyed error, which
can happen on server failover if we drop the Domains of the server that
went away, it would set the query to only go to Upstream from then on,
which meant:

1) Query performance would drop, for any queries that hit ViewDestroyed,
and that performance drop would have persisted until at least a cluster
was completely reset (I think)
2) `correct_data_after_restart` and `end_to_end_with_restart`, which
checked that we don't run into (1), were flaky if they hit the
ViewDestroyed error. This happened occasionally and depended on which
server was executing the query and which was the leader.

This commit makes it so that ViewDestroyed queries can be retried
through Noria and the tests are no longer flaky.

Additionally, this change adds logic to remove a read handle from the
local readers cache if its corresponding write handle was dropped. This
was initially suspected as the root cause of the flaky test, but turned
out not to be. Even so, this would at least mean extra, unnecessary
retries when we find out the reader is no longer usable, and perhaps
even cause issues similar to the actual underlying issue under certain
circumstances.

Other cleanup changes include:
- DRY out logic in readers.rs for getting a reader from the local/global
  reader caches. This is a bare fn rather than member fn due to borrow
checker lifetime contraints around partially borrowing `self`.
- Move `invalidate_prepared_cache_entry` from a static fn on `Backend`
  to a member of `PrepareResult` and rename to `make_upstream_only`.

Fixes: ENG-1383, ENG-1385, ENG-1395
Change-Id: I935f81997a3bc856fbb58288d71f6cf1d5b0831d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2701
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>