errors: Make constructor functions macros with location info
Currently we have two main ways of constructing the more "common" error
variants - we can construct-and-return them in one go via the
`internal!`, `unsupported!`, etc. macros, or we can construct them to be
returned directly (eg if inside of a `map_err` closure or similar place)
via the `internal_err`, `unsupported_err` etc. functions. The second
method, despite being the only option in the places where it is used,
has the big disadvantage that it omits location info in debug builds,
which can be *quite* nice when tracking down the original source of an
error.

This commit unifies the two approaches, bringing location info in debug
builds to the `internal_err` etc. constructor variants, by making them
into *macros*, which take the same arguments as eg `format!` (this also
has the advantage of removing one level of nesting at the callsite when
we need a format string!).

The diff here is pretty large, but almost all of it is
mechanical (literally, it was done with a find+replace and a regex) -
the relevant bits are all in `noria/errors/src/lib.rs`.

Fixes: ENG-410
Change-Id: I9dce830cae5c0660fe8b629d7cf911496ecf3cc3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2703
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>