server: Simplify + fix MIR -> dataflow join node lowering
The way the emit list was previously built during lowering from the MIR
join node to dataflow, if the column list in the MIR node was anything
other than literally the columns in the left-hand side followed by the
columns in the right-hand side (which it usually is, except in certain
cases where two joins are stacked on top of each other), we'd end up
with columns in a different order in the actual join dataflow node than
we had in the MIR node.  Then, when child nodes went to look up column
indices in the MIR node, those column indices wouldn't match the actual
columns returned by the dataflow node, causing all sorts of
issues (including, in this case, mismatched column types).

The actual implementation here was *way* more complicated than it needed
to be regardless - we can just iterate over the columns once and build
up the projected columns and join config as we go, which makes it
impossible for those lists to not match.

The way this is all modeled is pretty silly - we store things one way in
the MIR node, then convert it to another representation to pass into the
constructor for Join, then convert it to a *third* representation in
that constructor to actually put on a field of the Join node! I'd like
to refactor that all to just use one way of representing things at some
point, but for now I want to keep this diff smaller.

Fixes: ENG-867
Change-Id: Id69f1bf36edfcc1408d281e4407abbe5c6bb11bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2706
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>