Add optional ALWAYS flag to CREATE CACHE FROM
Adds optional ALWAYS flag to CREATE CACHE FROM. The idea of this
is to give users the ability to specify a query that they always want to
be served from ReadySet, regardless of what autocommit is set to.

This also updates SHOW CACHES to include a new column that indicates
whether this flag has been set or not.

This also updates the query status cache to record whether this flag was
set at time of migration or not.

Change-Id: Ia4142d2cf472af55433b7f1c28f9cea2bd08ad1a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2707
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>