noria-client: merge handle and execute select
Currently `handle_select` and `execute_prepared_statement`
are separate methods, that share most of the logic. It looks
like the fact that they are separate caused them to get out
of sync, so some logic differs between the two, when I believe
it shouldn't be the case. To avoid discrepancies like that in
the future, this CL merges the two into a single new method
called `execute_select`.

Change-Id: Ib1d302507f2acd8b058230db566d721b592f0d50
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2715
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>