dataflow: Rewrite evictions through generated columns
A node generating columns from itself means that those columns fully
depend on the input columns - which means that if we evict some keys
from the input columns, we should evict any matching keys from the
generated columns, including all the way down the graph - otherwise
we'll break the second partial law and get perma-stale result sets
further on down the graph. This is easily shown with a test case that
constructs a paginate node below a join that has to evict on writes -
that eviction just evicts the group from the paginate node, but none of
the keys for the individual pages either in the node or in the reader
below the paginate node in the graph.

This commit breaks out the bit of `trigger_downstream_evictions` that's
determining which replay paths to trigger along, and then extends it to
check if the key we just evicted from has any dependent generated
columns, and if so builds up a new set of keys to evict by selecting the
generated column indices from the rows we just evicted (which we now
need to return from both `evict_keys` and `evict_bytes`). Then, we use
those keys instead of the original set of keys to trigger a new series
of evictions down the replay path sourced at the generated columns.

Efficiently determining whether the columns we evicted from are used to
generate columns also required adding a little bit more information to
the DomainRequest::GeneratedColumns message that the domain gets sent to
notify it about generated columns - now, we don't just know the columns
that're generated, but also the columns they're generated from. This
also includes making `ColumnRef` generic on the type of node index we're
using (local or global), since within domains we generally prefer to use
local node indices wherever possible.

Fixes: ENG-1561
Change-Id: I740e4309cf7249eb7955b2e45c4c56762b1772fe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2720
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>