noria-client: handle LIMIT and OFFSET in reader as fallback
Currently topk and pagination support are experimental, and at
least for the topk node a lot of overhead is added for no performance
gains over simply returning the first k rows for a key/key set.

On the other hand LIMIT is already fully supported at the reader
level, since it has to be able to aggregate multiple keys with
a LIMIT clause anyway, and OFFSET can be easily supported by
skipping some rows at the reader level. Although skipping is
sub-optimal, at the very least it is still competetive with
MySQL and PostgreSQL that also have to skip results when OFFSET
is provided.

To this end, this change implements LIMIT and OFFSET as post
lookup filters. When server support for topk and pagination
is disabled, it will not send queries with LIMIT and OFFSET to
the server, and will instead strip them and send as lookup
parameters.

Changes in this CL:
Add an RPC call to check for topk+paginate support
Add an offset parameter to ResultIterator
Add limit/offset parameters to ViewQuery
Optionally remove LIMIT and OFFSET during rewrite phase
Send LIMIT and OFFSET during lookup *if* not enabled in readyset

Fixes: ENG-1573
Change-Id: Ia2a46236f0441f8e8148722fe8ab083ca774171f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2724
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>