docs: Create section and more docs for git
Add a large document including many references on the model of git and
how we use it at ReadySet.

Expand the tooling to add additional sections for other editors. This
could eventually be split into multiple files for different editors
and tools that people make use of.

This will be followed up with more documentation around Gerrit to
clarify that relationship too.

Change-Id: Ia3c2154ce071649b6820f1c20c290feb792ae110
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2728
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>