nom-sql: Remove Table::alias
Lots of places where we use the `nom_sql::Table` struct today aren't
places where it's valid to have an alias - things like the subject of a
`CREATE TABLE` , `DROP TABLE`, or `ALTER TABLE` statement, the `table`
field of `Column`, etc. In fact, the *only* place an alias is acceptable
is in the FROM clause of a query, either directly as a table or in a
join. This is annoying due to being an incorrect representation of the
syntax of the language, but perhaps more importantly prevents us from
effectively using `Table` throughout noria-server to refer to
schema-qualified tables, which we're now going to be supporting.

This commit removes `alias` from the Table struct entirely, instead
opting to wrap it in a `TableExpr` struct which has a `table` and an
`alias` field for the FROM clause of a SELECT statement, and then
basically just chases the type errors from there. This *slightly* paves
the road for a future world where we support more complex table
expressions (like actual expressions, or subqueries) directly in the
FROM clause - once we have those, we can make a `TableExprInner` enum
and make that the type of the `table` field in TableExpr.

Finally, there's lots of places in this commit where I'm leaving TODO
notes for myself to update things to care about schema - those will
quickly be addressed in follow-up commits.

Fixes: ENG-1585
Change-Id: I799ff2a21161a6efad62f351c1fdd9afd3ac5b74
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2734
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>