consensus: Implement standalone authority
When running in standalone mode there is little to gain
by requiring use of consul or zk, as no discovery takes
place. To this end this CL introduces a new kind of authority
that persists the schema to disk in a RocksDB database and
keeps other data ephemeral like local authority does.

Perhaps RocksDB is an overkill and we could do with
a binary file, however in theory authority may want to
store arbitrary keys in the future, and RocksDB works
well if it does.

Change-Id: Icdf08e40a31b23ddae50ca46d3dd93b2e5ba07ba
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2739
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>