sql-passes: Make all rewrite passes schema-aware
This commit goes through basically everything in `noria-sql-passes`, and
updates everywhere that only currently cares about the name of a table
to also care about that table's schema. This includes also updating
`ReadySetError::TableNotFound` to include a `schema` field, and making
the schema maps that're passed into schema-aware rewrite passes keyed on
Table instead of SqlIdentifier. To keep the commits small and focused,
the implied changes of that last bit are cut off for now inside of
noria-server, where we just construct `view_schemas` and `base_schemas`
from `self.view_schemas` and `self.base_schemas` respectively but
filling in None for all the schemas - next up will be a change to the
actual field types to key tables based on schema.

Refs: ENG-1580
Change-Id: I1e796a75707858cbdf350b51329eb375989cf73f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2743
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>