nom-sql: Make view names Tables, not SqlIdentifiers
VIEWs, like TABLEs, can live inside of schemas/dbs, and when created can
be qualified with the name of that schema or db in the same way that
tables can. This commit changes our AST to support that, by making the
`name` field of the CreateViewStatement a Table and the `names` field of
DropViewStatement a `Vec<Table>`, instead of just bare SqlIdentifiers.

Refs: ENG-1580
Change-Id: I111d8387ca613f55a4ab43a437a6728dcfb846e8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2745
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>