logictests: better visibility into tests
Currently we have no ability to tell which tests finished
succesfully if one or more of them hangs. This changes reports
the name of passing tests in addition to the names of the failing
tests, so it is easier to see which test is missing (and therfore)
is hanging. To make things even more obvious, it will report if
a logictest has been running for over 10 minutes, and will report
timing for all tests regardless of passing status.

Change-Id: Ic7e2acbc357c299a868052271fbbdb8febcb2c19
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2746
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>