consensus: add a method that lets optimize a clone away
Whenever the leader state is updated, the success status is
checked, but the updated state itself is ignored, however
some types of Authority, such as the LocalAuthority and the
incoming StandaloneAuthority can get better performance by
avoiding an extra clone to return the update states, which
is then dropped anyway.

This change adds a new trait that allows Authorities to
optimize for this use case by implementing the optimized
update_controller_in_place method. This speeds up nightly
builds tremendously.

Change-Id: Idfe5d08b3c600a3b21232dcf2895df979f6361e4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2748
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>