dataflow: Remove assert on duplicate columns in join output
I'm not sure why this was there in the first place, but it's definitely
possible for columns to appear multiple times in join output (in this
case, it's happening with EXISTS queries after 8377e5fc2 (server:
Simplify + fix MIR -> dataflow join node lowering, 2022-07-15). Things
appear to work perfectly fine without this check, so this commit just
removes it.

Change-Id: I56dca3bdee86c8288c70a917aee4be9d05d8ef7b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2749
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>