test: Removed alter table tests.
The re-evaluation of these tests happened because one of them was being
flaky and failing in CI, but I was unable to reproduce the failure
locally.
However, the ALTER TABLE tests should be removed (and we should also
remove the ALTER TABLE handling in the Recipe in the future), as the
solution for schema changes has changed, and Noria won't be handling
ALTER TABLE statements directly anymore.

Refs: ENG-1339
Change-Id: Id11bd3bfc34f751bf43d2ffe949bada042023a25
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2752
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>