sql-passes: Rewrite + rename negation removal pass
The negation removal pass originally only existed because the expression
evaluation in dataflow didn't support NOT anywhere - so we tried to
remove negation by negating negated expressions in-place (eg by changing
`NOT x = y` to `x != y`), returning *an error* in cases where an
expression wasn't statically negatable (eg a function call). Not only is
that no longer necessary (we support NOT in dataflow expression
evaluation now, and have for quite some time) it's actually *incorrect*
- MySQL allows NOT on any expression; it'll just coerce it to an integer
and then NOT that integer.

To make it clearer that this rewrite pass is really just a normalization
pass, this commit renames `negation_removal` to `normalize_negation`,
and updates it to be implemented using a Visitor instead of the manual
pattern matching it was using previously.

Change-Id: I321e319eaa2805d96feab86363c24f4337badc43
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2753
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>