sql: Get rid of ToFlowParts trait
Adding a query to the database via an extension trait on strings is a
little... odd, and has always thrown me for a loop when trying to trace
through it. Not to mention that it looks like now that we have
ChangeList this is actually only ever used for tests - via the
`add_query` method!

Now, instead, every test just uses `add_parsed_query`, parsing at the
callsite if necessary, which is a much cleaner way of doing things
that's more consistent with how the recipe changelist stuff works
anyway now. This also allowed deleting a whole load of extra related
stuff.

Refs: ENG-1594
Change-Id: I5e2ed6cf4320a3934f74f59f39992ea09b8ce309
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2757
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>