sql: Refactor the entrypoints to SqlIncorporator
Previously, the main entrypoint for the Recipe to add a new query to the
SqlIncorporator was the `add_parsed_query` function. This was passed a
`SqlQuery`, an optional name, an `is_leaf` boolean, and a mutable
reference to a Migration to add the nodes to. This was a particularly
awkward API for the way we were using it, primarily deriving from taking
a full `SqlQuery`:

* We actually only know how to handle views, tables, and select
  statements - all other `SqlQuery` variants return an error (even
  though we've already statically determined we're not going to have
  them because we're going through the ChangeList now!)
* The `name` argument is only relevant for newly added queries - in the
  case of tables and views it's obviously already in the statement (and
  in fact we assert that the provided name is equal to the one in the
  statement!). This is especially awkward for the upcoming support for
  multiple schemas.
* The `is_leaf` argument is always true for queries and views (but maybe
  shouldn't be true for views; more on that later) and always false for
  tables, which means if we already know what kind of statement we're
  adding it's completely useless.

This commit revamps that API by removing `add_parsed_query` and
replacing it with three functions: `add_table`, `add_view`, and
`add_query`, which are called from the relevant pattern matches in
`Recipe::activate`. The signatures of these functions are all as simple
as possible, including the return value, which since Recipe::activate
doesn't care no longer return `QueryFlowParts` - the only one that
actually returns a value is `add_query`, which returns the name of the
added query, useful in the case that the caller didn't supply one.

Since we aren't returning `QueryFlowParts` anymore, some assertions had
to go away in the tests in `sql/mod.rs` - that's a bit of a loss, but
I'd argue that basically all the tests in that file are too tighly
coupled to the implementation anyway, so this feels like a small loss in
comparison to the gain in simplicity for non-test code.

An earlier iteration of this change used `add_cache` instead of
`add_query` - I decided to go with query instead because this area of
the code tends to use that terminology for cached queries instead of the
newer "cache" terminology - if we want to, later we can refactor all of
the migration code to more consistently use "cache", but for now I
wanted to go with local consistency to avoid future confusion.

Fixes: ENG-1594
Change-Id: I463f356733168cb6711c31f64bee6f4e49001459
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2764
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>