controller: Don't lookup tables by node name
Currently, when looking up a table's node index as part of handling a
`/table_builder` RPC, we first look in the recipe, then, if we don't
find the table, we build up a BTreeMap of all tables keyed by the name
of the *dataflow* node in-place (self.tables()) and index into that.
This is behavior that's necessary for integration tests that add nodes
directly (eg `mig.add_base`) rather than through the SQL
interface (`extend_recipe`), since the former totally bypasses the
recipe. Setting aside that the actual lookup is pretty wildly
inefficient, it also breaks down as we want to add table schema in as
many places as possible - I don't think it'd be good to add an optional
`schema` field to every single dataflow node, for example.

Fortunately, since this is only ever actually used for noria-server
integration tests, there's a pretty easy workaround - `mig.add_base`
returns the index of the newly created node, so we can just return that
from the migration and then use that as the query to look up the table
rather than its name. This commit does that, adding a new
`table_by_index` method and corresponding RPC to Controller. We *could*
go a little further to add some type-based overloads to the API, but I
figured since this is only used for tests it was better to keep things
simple.

Change-Id: Ie72ecb0dd4dc08a508cd71affc1c1f4a1fb6e531
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2768
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>