Add Parsing of COM_SET_OPTION packet
Multiple php frameworks, as well as diesel send a COM_SET_OPTION packet
to enable multi-statement support. Currently we don't parse this packet
type, and as a result we never reply, leaving the connection hanging.

Readyset already supports receiving multiple semicolon separated
statements. We can therefore parse this packet, and simply reply ok
since this feature is already enabled in ReadySet.

Tested that this causes the Diesel framework test to pass, so I've
re-enabled pipeline testing on it with this commit.

Change-Id: If9d7ce87616df6f5516a60a53f47bb1afe21940c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2769
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Corbin Hall <sif@readyset.io>