noria-client: Fix error message for unsupported key operator
Previously any error encountered when converting a key and binary
operator for a ViewQuery to a KeyComparsion would be treated as an
"EmptyKey" error - even though the error *can* be caused by an invalid
binary operator when comparing against a parameter, eg `!=`. This fixes
that by making the conversion to KeyComparsion return a ReadySetError
and just using that as the return value.

Also, this changes the conversion to KeyComparison to used a named
function rather than a From impl, since it improves readability at the
call-site and allows easier go-to-definition.

Change-Id: I542329d6e056eacf4b9f68753cfad81bd17a7962
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2775
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>