Update reuse_similar_query test to work around unrelated bug
This test was intending to cover an issue, ENG-67, where equivalent but
not *syntactially* equivalent queries didn't properly get reused -  that
issue appears to have been fixed since, but we didn't realize it had
been fixed because it was hiding behind ENG-1612, which prevented us
from executing queries with comparisons like `NOT (x != ?)`. This commit
rewrites the `reuse_similar_query` test to use two still-equivalent
queries, but ones which we can successfully execute otherwise, and then
de-ignores it since it's passing now.

Also, noria-mysql dev-depends on `readyset-tracing` now, since it's
annoying to have to keep adding that dependency to get logging in tests

Change-Id: I43b4a79fe89a1aa7dfdc177437846c6d54ec3891
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2777
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>